Added DRA.cpp HAL for fast digital port writes and ADC reads to allow tighter sensor pulse timing. Refactored Sensors.cpp to use the new HAL.

+++ lib/Hardware/Sensors/Sensors.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <Arduino.h>
 
 /**
 * @brief Struct que contiene las lecturas de todos los sensores.
@@ -7,8 +8,8 @@ struct SensorData {
     int right;
     int center;
     int left;
-    int lineLeft;
-    int lineRight;
+    uint16_t lineLeft;
+    uint16_t lineRight;
 };
 
 /**
